requirements.txt
pillow
cssutils
